plotted std 5 rather than std 3 because there are 10 points on the std curve
However, after plotting it I saw that many of the values are already negative for std 5. It looks like the top std is always maxed out all the way back in original MFI, same with std 2 and even std 3 much of the time. Going back to std 3 for the levy jennings plot because it is mostly in the middle between zero and the max values. Also, the fact that many of the top points on the std curves for each sample have the same value is because of original max MFI, which carries forward to being the same for each sample after buffer normalization.